% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R, R/update.R
\name{update_registry}
\alias{update_registry}
\alias{sync}
\alias{update_submodules}
\title{Update the CRAN repo}
\usage{
update_registry(path = ".")

update_submodules(path = ".", skip = ".registry")
}
\arguments{
\item{path}{path to monorepo}
}
\description{
Synchronizes submodules in the cranrepo from a the global registry
}
