% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monorepos.R
\name{sync_from_registry}
\alias{sync_from_registry}
\title{Update the Monorepo}
\usage{
sync_from_registry(monorepo_url = Sys.getenv("MONOREPO_URL"))
}
\arguments{
\item{monorepo_url}{full git URL of monorepo}
}
\description{
Synchronizes submodules in the monorepo from a registry file.
}
